requirements.txt
h5py==2.10.0
importlib-metadata==0.23
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.17.3
opencv-contrib-python==4.1.1.26
requests==2.22.0
scikit-learn==0.21.3
tensorboard==2.0.0
tensorflow==2.0.0

